Move render fields into separate component
